Modified the buildnt.cmd to always clean before a build.
git-svn-id: svn://svn.code.sf.net/p/eapidk/code/trunk@57 3723b92f-f4ce-431d-9be9-3c706c148987

+++ apps/common/EApiAHStr.c
@@ -159,7 +159,7 @@ EApiAHGetString(
 
   EApiStatusCode=EApiBoardGetString(StringID, pString, &StringLenLcl);
   if(EAPI_STATUS_TEST_NOK(EApiStatusCode))
-      StringLenLcl=(uint32_t)EApiAHCreateErrorString(EApiStatusCode, pString, StringBufferLenBck);
+      StringLenLcl=(uint32_t)EApiAHCreateErrorString((uint32_t)EApiStatusCode, pString, StringBufferLenBck);
 #if (STRICT_VALIDATION>0)
   EAPI_APP_RETURN_ERROR_IF(
       EApiAHGetStringAlloc,
